refactor(auth): Replace AuthProvider with useAuth hook across components for improved authentication management
