+++ frontend/src/app/workspaces/layout.tsx
@@ -12,8 +12,8 @@ import { DynamicNavbar } from "@/components/nav/dynamic-nav"
 import { AppSidebar } from "@/components/sidebar/app-sidebar"
 import { Button } from "@/components/ui/button"
 import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar"
+import { useAuthActions } from "@/hooks/use-auth"
 import { useWorkspaceManager } from "@/lib/hooks"
-import { useAuth } from "@/providers/auth"
 import { WorkflowBuilderProvider } from "@/providers/builder"
 import { WorkflowProvider } from "@/providers/workflow"
 import { WorkspaceIdProvider } from "@/providers/workspace-id"
@@ -118,7 +118,7 @@ function WorkflowView({
 }
 
 function NoWorkspaces() {
-  const { logout } = useAuth()
+  const { logout } = useAuthActions()
   const handleLogout = async () => {
     await logout()
   }
