+++ frontend/src/components/cases/case-table.tsx
@@ -16,10 +16,10 @@ import { DataTable, type DataTableToolbarProps } from "@/components/data-table"
 import { TooltipProvider } from "@/components/ui/tooltip"
 import { useToast } from "@/components/ui/use-toast"
 import { useCasesPagination } from "@/hooks"
+import { useAuth } from "@/hooks/use-auth"
 import { useWorkspaceDetails } from "@/hooks/use-workspace"
 import { getDisplayName } from "@/lib/auth"
 import { useDeleteCase } from "@/lib/hooks"
-import { useAuth } from "@/providers/auth"
 import { useWorkspaceId } from "@/providers/workspace-id"
 
 export default function CaseTable() {
