+++ frontend/src/components/nav/user-nav.tsx
@@ -25,11 +25,12 @@ import {
 import UserAvatar from "@/components/user-avatar"
 import { siteConfig } from "@/config/site"
 import { userDefaults } from "@/config/user"
+import { useAuth, useAuthActions } from "@/hooks/use-auth"
 import { useWorkspaceManager } from "@/lib/hooks"
-import { useAuth } from "@/providers/auth"
 
 export default function UserNav() {
-  const { user, logout } = useAuth()
+  const { user } = useAuth()
+  const { logout } = useAuthActions()
   const { clearLastWorkspaceId } = useWorkspaceManager()
   const params = useParams<{ workspaceId?: string }>()
   const workspaceId = params?.workspaceId
